show t and recalc

+++ Van.Core/Methods/Exponential.cs
@@ -8,18 +8,19 @@ namespace Van.Core.Methods
 {
     public class Exponential
     {
-        public Exponential(List<double> t, List<double> delta)
+        public Exponential(List<int> t, List<int> delta)
         {
             this.t = t;
             this.delta = delta;
         }
 
-        public List<double> t = new List<double>();
-        public List<double> delta = new List<double>();
+        public List<int> t = new List<int>();
+        public List<int> delta = new List<int>();
         public double r => delta.Where(x => x == 1).Count();
         public double n => t.Count();
 
-        public double lambda() {
+        public double lambda()
+        {
             return r * Math.Pow(t.Sum(), -1);
         }
