+++ Van.Core/Methods/Gompertz.cs
@@ -8,7 +8,7 @@ namespace Van.Core.Methods
 {
     public class Gompertz
     {
-        public Gompertz(List<double> t, List<double> delta, double b, double epsilon, double? a = null)
+        public Gompertz(List<int> t, List<int> delta, double b, double epsilon, double? a = null)
         {
             this.t = t;
             this.delta = delta;
@@ -17,8 +17,8 @@ namespace Van.Core.Methods
             this.a = a != null ? (double)a : epsilon;
         }
 
-        public List<double> t = new List<double>();
-        public List<double> delta = new List<double>();
+        public List<int> t = new List<int>();
+        public List<int> delta = new List<int>();
         public double r => delta.Where(x => x == 1).Count();
         public double n => t.Count();
         public double b { get; set; }
@@ -26,7 +26,8 @@ namespace Van.Core.Methods
         public double a { get; set; }
 
 
-        public double FirstSum(double x) {
+        public double FirstSum(double x)
+        {
             double firstSum = 0;
 
             for (int i = 0; i < n; i++)
@@ -58,10 +59,11 @@ namespace Van.Core.Methods
                 thirdSum += Math.Exp(t[i] * x);
             }
 
-            return n + (1-x*x)*thirdSum;
+            return n + (1 - x * x) * thirdSum;
         }
 
-        public double function(double x) {
+        public double function(double x)
+        {
             return FirstSum(x) - SecondSum(x) * ThirdSum(x);
         }
