+++ Van.Core/Methods/Relay.cs
@@ -8,18 +8,19 @@ namespace Van.Core.Methods
 {
     public class Relay
     {
-        public Relay(List<double> t, List<double> delta)
+        public Relay(List<int> t, List<int> delta)
         {
             this.t = t;
             this.delta = delta;
         }
 
-        public List<double> t = new List<double>();
-        public List<double> delta = new List<double>();
+        public List<int> t = new List<int>();
+        public List<int> delta = new List<int>();
         public double r => delta.Where(x => x == 1).Count();
         public double n => t.Count();
 
-        public double FirstSum() { 
+        public double FirstSum()
+        {
             double sum = 0;
 
             for (int i = 0; i < n; i++)
@@ -30,7 +31,8 @@ namespace Van.Core.Methods
             return sum;
         }
 
-        public double lambda() {
+        public double lambda()
+        {
             return Math.Pow(1 / (2 * r) * FirstSum(), -2);
         }
